# server/swagger.yaml
openapi: 3.0.0
info:
  title: RapiDevKit
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /items/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    post:
      summary: CRUD ACTION update
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
  /items/findByActive/{key}:
    get:
      summary: CRUD ACTION findByActive
      tags:
        - Item
      parameters:
        - name: key
          in: path
          description: Id of the resource Active to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/findByCreated/{key}:
    get:
      summary: CRUD ACTION findByCreated
      tags:
        - Item
      parameters:
        - name: by
          in: path
          description: start date
          required: true
          schema:
            type: integer
        - name: to
          in: path
          description: end date
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/findByCreator/{key}:
    get:
      summary: CRUD ACTION findByCreator
      tags:
        - Item
      parameters:
        - name: key
          in: path
          description: Id of the resource Creator to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/findByProtected/{key}:
    get:
      summary: CRUD ACTION findByProtected
      tags:
        - Item
      parameters:
        - name: key
          in: path
          description: Id of the resource Protected to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/findByType/{key}:
    get:
      summary: CRUD ACTION findByType
      tags:
        - Item
      parameters:
        - name: key
          in: path
          description: Id della risorsa Type da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/{id}/getLink:
    get:
      summary: CRUD ACTION getLink
      tags:
        - Item
      parameters:
        - name: id
          in: path
          description: ID of Item from Link
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
    
  #-----  ItemType  -----
...
  /types/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemType"
    post:
      summary: CRUD ACTION update
      tags:
        - ItemType
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemType"
  /types/findByName/{key}:
    get:
      summary: CRUD ACTION findByName
      tags:
        - ItemType
      parameters:
        - name: key
          in: path
          description: Id of the resource Name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemType"
    
  #-----  Role  -----
...
  /roles/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    post:
      summary: CRUD ACTION update
      tags:
        - Role
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Item:
      properties:
        _id:
          type: string
        Active:
          type: boolean
        Created:
          type: integer
        Data:
          type: string
        Modified:
          type: integer
        Order:
          type: integer
        Protected:
          type: boolean
        Status:
          type: string
      required:
        -  _id
        - Active
    ItemType:
      properties:
        _id:
          type: string
        Name:
          type: string
        Schema:
          type: string
        Validation:
          type: string
        Visualization:
          type: string
      required:
        -  _id
    Role:
      properties:
        _id:
          type: string
        Name:
          type: string
      required:
        -  _id
...